requirements.txt
Django==4.1.3
environs==9.5.0
phonenumbers==8.13.0
django-phonenumber-field==7.0.0